refactor (CardViewUtil): 확장함수 Util 패키지로 독립
