+++ src/main/kotlin/blackjack/view/OutputView.kt
@@ -1,8 +1,6 @@
 package blackjack.view
 
 import blackjack.domain.blackjack.BlackJack
-import blackjack.domain.card.CardMark
-import blackjack.domain.card.CardValue
 import blackjack.domain.participants.User
 import blackjack.domain.result.BlackJackResult
 import blackjack.domain.result.BlackJackResultUser
@@ -26,14 +24,13 @@ class OutputView {
     }
 
     fun outputCard(user: User) {
-        val cardText = user.cards.toList().joinToString(", ") { it.value.pattern() + it.mark.name() }
+        val cardText = user.cards.toList().joinToString(", ") { it.toText() }
         print("\n${user.name}카드: $cardText")
     }
-
     fun outputDealerDraw() { println("\n\n딜러는 16이하라 한장의 카드를 더 받았습니다.") }
 
     private fun outputCardForDealer(user: User) {
-        val cardText = user.cards.toList()[0].let { it.value.pattern() + it.mark.name() }
+        val cardText = user.cards.toList()[0].toText()
         print("\n${user.name}카드: $cardText")
     }
 
@@ -48,21 +45,4 @@ class OutputView {
     }
 
     private fun outputOutcome(result: BlackJackResultUser) { println("${result.name}: ${result.rateOfReturn}") }
-
-    private fun CardMark.name(): String =
-        when (this) {
-            CardMark.CLOVER -> "클로버"
-            CardMark.DIA -> "다이아몬드"
-            CardMark.HEART -> "하트"
-            CardMark.SPADE -> "스페이드"
-        }
-
-    private fun CardValue.pattern(): String =
-        when (this) {
-            CardValue.ACE -> "A"
-            CardValue.KING -> "K"
-            CardValue.QUEEN -> "Q"
-            CardValue.JACK -> "J"
-            else -> this.value.toString()
-        }
 }
